refactor: 사다리 관련 객체들을 별도의 패키지로 이동
